# api-catalogo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio de Catálogo de Homelab Collector Platform.",
        "title": "Homelab Collector Platform - Catalogo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9094",
    "basePath": "/",
    "paths": {
...
        "/api/v1/catalogo/tipos-sensor": {
            "get": {
                "description": "Obtiene la lista de todos los tipos de sensor registrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos de Sensor"
                ],
                "summary": "Listar tipos de sensor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TipoSensor"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo tipo de sensor en el catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos de Sensor"
                ],
                "summary": "Crear tipo de sensor",
                "parameters": [
                    {
                        "description": "Datos para crear el tipo de sensor",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTipoSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TipoSensor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/catalogo/tipos-sensor/{id}": {
            "get": {
                "description": "Obtiene los detalles de un tipo de sensor por su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos de Sensor"
                ],
                "summary": "Obtener tipo de sensor por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tipo de sensor (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TipoSensor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza los datos de un tipo de sensor existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos de Sensor"
                ],
                "summary": "Actualizar tipo de sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tipo de sensor (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateTipoSensorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TipoSensor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un tipo de sensor del catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tipos de Sensor"
                ],
                "summary": "Eliminar tipo de sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del tipo de sensor (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.CreateTipoSensorRequest": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "unidad_medida": {
                    "type": "string"
                }
            }
        },
...
        "model.TipoSensor": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "unidad_medida": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateTipoSensorRequest": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "unidad_medida": {
                    "type": "string"
                }
            }
        },
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}